requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu
Flask==3.0.2
flask-cors==4.0.0
opencv-python-headless==4.9.0.80
torch==2.2.1+cpu
torchvision==0.17.1+cpu
Pillow==10.2.0
numpy==1.26.4